[front] Remove Tailwind validation logic from server actions and relocate to client executable logic
