+++ front/lib/api/files/client_executable.ts
@@ -1,4 +1,3 @@
-import { validateTailwindCode } from "@app/lib/actions/mcp_internal_actions/servers/common/viz/validation";
 import { getFileContent } from "@app/lib/api/files/utils";
 import type { Authenticator } from "@app/lib/auth";
 import { FileResource } from "@app/lib/resources/file_resource";
@@ -12,6 +11,52 @@ import {
   normalizeError,
   Ok,
 } from "@app/types";
+
+// Regular expression to capture the value inside a className attribute. This pattern assumes
+// double quotes for simplicity.
+const classNameRegex = /className\s*=\s*"([^"]*)"/g;
+
+// Regular expression to capture Tailwind arbitrary values:
+// Matches a word boundary, then one or more lowercase letters or hyphens,
+// followed by a dash, an opening bracket, one or more non-']' characters, and a closing bracket.
+const arbitraryRegex = /\b[a-z-]+-\[[^\]]+\]/g;
+
+/**
+ * Validates that the generated code doesn't contain Tailwind arbitrary values.
+ *
+ * Arbitrary values like h-[600px], w-[800px], bg-[#ff0000] cause visualization failures
+ * because they're not included in our pre-built CSS. This validation fails fast with
+ * a clear error message that gets exposed to the user, allowing them to retry which
+ * provides the error details to the model for correction.
+ */
+function validateTailwindCode(code: string): Result<undefined, Error> {
+  const matches: string[] = [];
+  let classMatch: RegExpExecArray | null = null;
+
+  // Iterate through all occurrences of the className attribute in the code.
+  while ((classMatch = classNameRegex.exec(code)) !== null) {
+    const classContent = classMatch[1];
+    if (classContent) {
+      // Find all matching arbitrary values within the class attribute's value.
+      const arbitraryMatches = classContent.match(arbitraryRegex) || [];
+      matches.push(...arbitraryMatches);
+    }
+  }
+
+  // If we found any, remove duplicates and throw an error with up to three examples.
+  if (matches.length > 0) {
+    const uniqueMatches = Array.from(new Set(matches));
+    const examples = uniqueMatches.slice(0, 3).join(", ");
+    return new Err(
+      new Error(
+        `Forbidden Tailwind arbitrary values detected: ${examples}. ` +
+          `Arbitrary values like h-[600px], w-[800px], bg-[#ff0000] are not allowed. ` +
+          `Use predefined classes like h-96, w-full, bg-red-500 instead, or use the style prop for specific values.`
+      )
+    );
+  }
+  return new Ok(undefined);
+}
 
 export async function createClientExecutableFile(
   auth: Authenticator,
